bot/requirements.txt
aiogram==2.25.1
psycopg2-binary
redis